Remove get_by_natural_key() again from user manager
Django really doesn't like using anything but the username field for a user's natural key.

In django/contrib/auth/backends.py:
```
    def authenticate(self, request, username=None, password=None, **kwargs):
        if username is None:
            username = kwargs.get(UserModel.USERNAME_FIELD)
        if username is None or password is None:
            return
        try:
            user = UserModel._default_manager.get_by_natural_key(username)
        except UserModel.DoesNotExist:
            # Run the default password hasher once to reduce the timing
            # difference between an existing and a nonexistent user (#20760).
            UserModel().set_password(password)
        else:
            if user.check_password(password) and self.user_can_authenticate(user):
                return user
```

Note that if `get_by_natural_key()` is there, it will call it with the
supplied username (which in our case is an email address) as argument.
This causes an error because our implementation assumes that the natural
key is a UUID.

In some other place, Django seems to assume that `get_username()` returns
the content of the `USERNAME_FIELD`, which we do not, and I encountered
some Django confusion in some other place because of that. We need to
fix this mess at some point.