[README BEFORE DEPLOYING] Start using datasets from kausal_common
Before deploying this commit, please deploy the previous one and
migrate the datasets app.

That will remove all the tables for the app.

After that, migrate the datasets app again with:

  manage.py migrate datasets zero --fake

That will remove the migration state to be ready for the
new datasets app with the same name.

Only after that deploy to this commit and migrate normally.