Expose values for all years in DashboardCardBlock
This adds a field `goalValues`, which contains all year-value pairs for
the specified goal.

We deprecate the old `goalValue`, which returned only the value that
belongs to the target year. This caused some problems for a customer
that had goal without a value for the target year, while having a value
for some other year.